+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ObjectsPage/Tabs/TabUseDataset.tsx
@@ -6,7 +6,7 @@ import {parseRef} from '../../../../../../../react';
 import {abbreviateRef} from '../../../../../../../util/refs';
 import {CopyableText} from '../../../../../../CopyableText';
 import {DocLink} from '../../common/Links';
-import {TabUseBanner} from '../../TabUseBanner';
+import {TabUseBanner} from '../../common/TabUseBanner';
 import {
   OBJECT_ATTR_EDGE_NAME,
   TABLE_ID_EDGE_NAME,
