+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/OpsPage/Tabs/TabUseOp.tsx
@@ -5,7 +5,7 @@ import {isValidVarName} from '../../../../../../../core/util/var';
 import {abbreviateRef} from '../../../../../../../util/refs';
 import {CopyableText} from '../../../../../../CopyableText';
 import {DocLink} from '../../common/Links';
-import {TabUseBanner} from '../../TabUseBanner';
+import {TabUseBanner} from '../../common/TabUseBanner';
 
 type TabUseOpProps = {
   name: string;
